refactor: move the self filter out of perception into manipulation
It takes a manipulation RobotModelConfig, so it is not the general self-filter
perception would want. Parked next to the planning utils that already handle
robot geometry until perception has a model-agnostic design for this.

Also dropped a mask test the previous-position test already covers.
