+++ dimos/manipulation/planning/utils/point_cloud_self_filter.py
